Add typescript typings
This will allow typescript users of this package to get full type
information as well as autocompletion should they use such a tool.

There are no actual code changes with this commit and should have no
impact on the tests themselves.

Note: Ideally, `@types/node` should have been a dependency in this
package as the `InspectOptions` type is defined within this package.
However, as that was the only type that was used, I decided to copy it
from the type package instead. The downside of this is that if
`InspectOptions` is updated, the index.d.ts won't automatically be
updated.

Another note: At the bottom of the index.d.ts, it exposes four different
versions of the test function. However, these are only a few of the
different versions that can be used as the arguments can be sent to it
in any order.